+++ npm/iptm/iptm_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-container-networking/npm/metrics"
-
+	"github.com/Azure/azure-container-networking/npm/metrics/promutil"
 	"github.com/Azure/azure-container-networking/npm/util"
 )
 
@@ -150,15 +150,15 @@ func TestAdd(t *testing.T) {
 		},
 	}
 
-	gaugeVal, err1 := metrics.GetValue(metrics.NumIPTableRules)
-	countVal, err2 := metrics.GetCountValue(metrics.AddIPTableRuleExecTime)
+	gaugeVal, err1 := promutil.GetValue(metrics.NumIPTableRules)
+	countVal, err2 := promutil.GetCountValue(metrics.AddIPTableRuleExecTime)
 
 	if err := iptMgr.Add(entry); err != nil {
 		t.Errorf("TestAdd failed @ iptMgr.Add")
 	}
 
-	newGaugeVal, err3 := metrics.GetValue(metrics.NumIPTableRules)
-	newCountVal, err4 := metrics.GetCountValue(metrics.AddIPTableRuleExecTime)
+	newGaugeVal, err3 := promutil.GetValue(metrics.NumIPTableRules)
+	newCountVal, err4 := promutil.GetCountValue(metrics.AddIPTableRuleExecTime)
 	metrics.NotifyIfErrors(t, err1, err2, err3, err4)
 	if newGaugeVal != gaugeVal+1 {
 		t.Errorf("Change in iptable rule number didn't register in prometheus")
@@ -191,13 +191,13 @@ func TestDelete(t *testing.T) {
 		t.Errorf("TestDelete failed @ iptMgr.Add")
 	}
 
-	gaugeVal, err1 := metrics.GetValue(metrics.NumIPTableRules)
+	gaugeVal, err1 := promutil.GetValue(metrics.NumIPTableRules)
 
 	if err := iptMgr.Delete(entry); err != nil {
 		t.Errorf("TestDelete failed @ iptMgr.Delete")
 	}
 
-	newGaugeVal, err2 := metrics.GetValue(metrics.NumIPTableRules)
+	newGaugeVal, err2 := promutil.GetValue(metrics.NumIPTableRules)
 	metrics.NotifyIfErrors(t, err1, err2)
 	if newGaugeVal != gaugeVal-1 {
 		t.Errorf("Change in iptable rule number didn't register in prometheus")
