+++ npm/metrics/promutil/util.go
@@ -1,4 +1,4 @@
-package util
+package promutil
 
 import (
 	"fmt"
