+++ npm/nwpolicy_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/Azure/azure-container-networking/npm/ipsm"
 	"github.com/Azure/azure-container-networking/npm/iptm"
 	"github.com/Azure/azure-container-networking/npm/metrics"
+	"github.com/Azure/azure-container-networking/npm/metrics/promutil"
 	"github.com/Azure/azure-container-networking/npm/util"
 
 	corev1 "k8s.io/api/core/v1"
@@ -92,8 +93,8 @@ func TestAddNetworkPolicy(t *testing.T) {
 		},
 	}
 
-	gaugeVal, err1 := metrics.GetValue(metrics.NumPolicies)
-	countVal, err2 := metrics.GetCountValue(metrics.AddPolicyExecTime)
+	gaugeVal, err1 := promutil.GetValue(metrics.NumPolicies)
+	countVal, err2 := promutil.GetCountValue(metrics.AddPolicyExecTime)
 
 	npMgr.Lock()
 	if err := npMgr.AddNetworkPolicy(allowIngress); err != nil {
@@ -131,8 +132,8 @@ func TestAddNetworkPolicy(t *testing.T) {
 	}
 	npMgr.Unlock()
 
-	newGaugeVal, err3 := metrics.GetValue(metrics.NumPolicies)
-	newCountVal, err4 := metrics.GetCountValue(metrics.AddPolicyExecTime)
+	newGaugeVal, err3 := promutil.GetValue(metrics.NumPolicies)
+	newCountVal, err4 := promutil.GetCountValue(metrics.AddPolicyExecTime)
 	metrics.NotifyIfErrors(t, err1, err2, err3, err4)
 	if newGaugeVal != gaugeVal+2 {
 		t.Errorf("Change in policy number didn't register in prometheus")
@@ -336,14 +337,14 @@ func TestDeleteNetworkPolicy(t *testing.T) {
 		t.Errorf("TestAddNetworkPolicy failed @ AddNetworkPolicy")
 	}
 
-	gaugeVal, err1 := metrics.GetValue(metrics.NumPolicies)
+	gaugeVal, err1 := promutil.GetValue(metrics.NumPolicies)
 
 	if err := npMgr.DeleteNetworkPolicy(allow); err != nil {
 		t.Errorf("TestDeleteNetworkPolicy failed @ DeleteNetworkPolicy")
 	}
 	npMgr.Unlock()
 
-	newGaugeVal, err2 := metrics.GetValue(metrics.NumPolicies)
+	newGaugeVal, err2 := promutil.GetValue(metrics.NumPolicies)
 	metrics.NotifyIfErrors(t, err1, err2)
 	if newGaugeVal != gaugeVal-1 {
 		t.Errorf("Change in policy number didn't register in prometheus")
